# perf: add initial benchmark