--- a/internal/testing/assert/assert.go
+++ b/internal/testing/assert/assert.go
@@ -14,7 +14,7 @@ import (
 )
 
 // Equal asserts that two values are equal.
-func Equal[T comparable](t *testing.T, got, want T, msg string, arg ...any) {
+func Equal[T comparable](t testing.TB, got, want T, msg string, arg ...any) {
 	t.Helper()
 	if got != want {
 		if msg == "" {
@@ -26,7 +26,7 @@ func Equal[T comparable](t *testing.T, got, want T, msg string, arg ...any) {
 }
 
 // DeepEqual asserts that two values are deeply equal.
-func DeepEqual[T any](t *testing.T, got, want T, msg string, arg ...any) {
+func DeepEqual[T any](t testing.TB, got, want T, msg string, arg ...any) {
 	t.Helper()
 	if !reflect.DeepEqual(got, want) {
 		if msg == "" {
@@ -38,7 +38,7 @@ func DeepEqual[T any](t *testing.T, got, want T, msg string, arg ...any) {
 }
 
 // NilError asserts that an error is nil.
-func NilError(t *testing.T, err error) {
+func NilError(t testing.TB, err error) {
 	t.Helper()
 	if err != nil {
 		t.Fatalf("expected nil error, got %q (%T)", err, err)
@@ -46,14 +46,14 @@ func NilError(t *testing.T, err error) {
 }
 
 // Error asserts that an error is not nil.
-func Error(t *testing.T, got, expected error) {
+func Error(t testing.TB, got, expected error) {
 	t.Helper()
 	if !errorsMatch(t, got, expected) {
 		t.Fatalf("expected error %q, got %v (%T vs %T)", expected, got, expected, got)
 	}
 }
 
-func errorsMatch(t *testing.T, got, expected error) bool {
+func errorsMatch(t testing.TB, got, expected error) bool {
 	t.Helper()
 	switch {
 	case got == expected:
@@ -68,7 +68,7 @@ func errorsMatch(t *testing.T, got, expected error) bool {
 }
 
 // StatusCode asserts that a response has a specific status code.
-func StatusCode(t *testing.T, resp *http.Response, code int) {
+func StatusCode(t testing.TB, resp *http.Response, code int) {
 	t.Helper()
 	if resp.StatusCode != code {
 		t.Fatalf("expected status code %d, got %d", code, resp.StatusCode)
@@ -87,7 +87,7 @@ func isSafeContentType(ct string) bool {
 }
 
 // Header asserts that a header key has a specific value in a response.
-func Header(t *testing.T, resp *http.Response, key, want string) {
+func Header(t testing.TB, resp *http.Response, key, want string) {
 	t.Helper()
 	got := resp.Header.Get(key)
 	if want != got {
@@ -97,13 +97,13 @@ func Header(t *testing.T, resp *http.Response, key, want string) {
 
 // ContentType asserts that a response has a specific Content-Type header
 // value.
-func ContentType(t *testing.T, resp *http.Response, contentType string) {
+func ContentType(t testing.TB, resp *http.Response, contentType string) {
 	t.Helper()
 	Header(t, resp, "Content-Type", contentType)
 }
 
 // Contains asserts that needle is found in the given string.
-func Contains(t *testing.T, s string, needle string, description string) {
+func Contains(t testing.TB, s string, needle string, description string) {
 	t.Helper()
 	if !strings.Contains(s, needle) {
 		t.Fatalf("expected string %q in %s %q", needle, description, s)
@@ -111,28 +111,28 @@ func Contains(t *testing.T, s string, needle string, description string) {
 }
 
 // BodyContains asserts that a response body contains a specific substring.
-func BodyContains(t *testing.T, resp *http.Response, needle string) {
+func BodyContains(t testing.TB, resp *http.Response, needle string) {
 	t.Helper()
 	body := must.ReadAll(t, resp.Body)
 	Contains(t, body, needle, "body")
 }
 
 // BodyEquals asserts that a response body is equal to a specific string.
-func BodyEquals(t *testing.T, resp *http.Response, want string) {
+func BodyEquals(t testing.TB, resp *http.Response, want string) {
 	t.Helper()
 	got := must.ReadAll(t, resp.Body)
 	Equal(t, got, want, "incorrect response body")
 }
 
 // BodySize asserts that a response body is a specific size.
-func BodySize(t *testing.T, resp *http.Response, want int) {
+func BodySize(t testing.TB, resp *http.Response, want int) {
 	t.Helper()
 	got := must.ReadAll(t, resp.Body)
 	Equal(t, len(got), want, "incorrect response body size")
 }
 
 // DurationRange asserts that a duration is within a specific range.
-func DurationRange(t *testing.T, got, minVal, maxVal time.Duration) {
+func DurationRange(t testing.TB, got, minVal, maxVal time.Duration) {
 	t.Helper()
 	if got < minVal || got > maxVal {
 		t.Fatalf("expected duration between %s and %s, got %s", minVal, maxVal, got)
@@ -144,7 +144,7 @@ type number interface {
 }
 
 // RoughlyEqual asserts that a numeric value is within a certain tolerance.
-func RoughlyEqual[T number](t *testing.T, got, want T, epsilon T) {
+func RoughlyEqual[T number](t testing.TB, got, want T, epsilon T) {
 	t.Helper()
 	if got < want-epsilon || got > want+epsilon {
 		t.Fatalf("expected value between %v and %v, got %v", want-epsilon, want+epsilon, got)