Results from my laptop:

```
$ make bench
go test -bench=. -benchmem
goos: darwin
goarch: arm64
pkg: github.com/mccutchen/websocket
cpu: Apple M2
BenchmarkReadFrame/64b-8                11431972               101.1 ns/op           120 B/op          4 allocs/op
BenchmarkReadFrame/256b-8                5405172               222.2 ns/op           312 B/op          5 allocs/op
BenchmarkReadFrame/1024b-8               1801424               667.1 ns/op          1080 B/op          5 allocs/op
BenchmarkReadFrame/65536b-8                29114             41176 ns/op           65601 B/op          5 allocs/op
BenchmarkReadFrame/1048576b-8               1804            664341 ns/op         1048655 B/op          5 allocs/op
PASS
ok      github.com/mccutchen/websocket  7.914s
```

More thorough benchmarks will likely require additional refactoring.